# satellite/overlay: avoid large statement for piece counts