--- a/satellite/satellitedb/overlaycache.go
+++ b/satellite/satellitedb/overlaycache.go
@@ -8,6 +8,7 @@ import (
 	"database/sql"
 	"encoding/hex"
 	"fmt"
+	"sort"
 	"strings"
 	"time"
 
@@ -930,49 +931,61 @@ func (cache *overlaycache) AllPieceCounts(ctx context.Context) (_ map[storj.Node
 }
 
 func (cache *overlaycache) UpdatePieceCounts(ctx context.Context, pieceCounts map[storj.NodeID]int) (err error) {
-	// TODO: add monkit stuff
 	defer mon.Task()(&ctx)(&err)
-
 	if len(pieceCounts) == 0 {
 		return nil
 	}
 
-	switch cache.db.Driver().(type) {
+	type NodeCount struct {
+		ID    storj.NodeID
+		Count int64
+	}
+	var counts []NodeCount
+
+	for nodeid, count := range pieceCounts {
+		counts = append(counts, NodeCount{
+			ID:    nodeid,
+			Count: int64(count),
+		})
+	}
+
+	sort.Slice(counts, func(i, k int) bool {
+		return counts[i].ID.Less(counts[k].ID)
+	})
+
+	switch t := cache.db.Driver().(type) {
+	case *sqlite3.SQLiteDriver:
+		err = cache.db.WithTx(ctx, func(ctx context.Context, tx *dbx.Tx) error {
+			query := tx.Rebind(`UPDATE nodes SET piece_count = ? WHERE id = ?`)
+			for _, count := range counts {
+				_, err := tx.Tx.ExecContext(ctx, query, count.Count, count.ID)
+				if err != nil {
+					return Error.Wrap(err)
+				}
+			}
+			return nil
+		})
 	case *pq.Driver:
-		return Error.Wrap(cache.postgresUpdatePieceCounts(ctx, pieceCounts))
+		var nodeIDs []storj.NodeID
+		var countNumbers []int64
+		for _, count := range counts {
+			nodeIDs = append(nodeIDs, count.ID)
+			countNumbers = append(countNumbers, count.Count)
+		}
+
+		_, err = cache.db.ExecContext(ctx, `
+			UPDATE nodes
+				SET piece_count = update.count
+			FROM (
+				SELECT unnest($1::bytea[]) as id, unnest($2::bigint[]) as count
+			) as update
+			WHERE nodes.id = update.id
+		`, postgresNodeIDList(nodeIDs), pq.Array(countNumbers))
 	default:
-		return Error.Wrap(cache.sqliteUpdatePieceCounts(ctx, pieceCounts))
+		return Error.New("Unsupported database %t", t)
 	}
-}
 
-func (cache *overlaycache) postgresUpdatePieceCounts(ctx context.Context, pieceCounts map[storj.NodeID]int) (err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	sqlQuery := `UPDATE nodes SET piece_count = newvals.piece_count FROM ( VALUES `
-	args := make([]interface{}, 0, len(pieceCounts)*2)
-	for nodeID, pieceCount := range pieceCounts {
-		sqlQuery += `(?::BYTEA, ?::BIGINT), `
-		args = append(args, nodeID, pieceCount)
-	}
-	sqlQuery = sqlQuery[:len(sqlQuery)-2]
-	// trim off the last comma+space
-	sqlQuery += `) newvals(nodeid, piece_count) WHERE nodes.id = newvals.nodeid`
-	_, err = cache.db.DB.ExecContext(ctx, cache.db.Rebind(sqlQuery), args...)
-	return err
-}
-
-func (cache *overlaycache) sqliteUpdatePieceCounts(ctx context.Context, pieceCounts map[storj.NodeID]int) (err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	var args []interface{}
-	sqlQuery := ""
-	updateSQL := "UPDATE nodes SET ( piece_count ) = ( ? ) WHERE id == ?;"
-	for nodeID, pieceCount := range pieceCounts {
-		sqlQuery += updateSQL
-		args = append(args, pieceCount, nodeID)
-	}
-	_, err = cache.db.DB.ExecContext(ctx, cache.db.Rebind(sqlQuery), args...)
-	return err
+	return Error.Wrap(err)
 }
 
 func convertDBNode(ctx context.Context, info *dbx.Node) (_ *overlay.NodeDossier, err error) {