What: Use array arguments instead of constructing a large single query.

Why: Concatenated SQL statements are harder to audit and often can end up slower.

```
Benchmark                                      old ns/op      new ns/op     delta
BenchmarkDB_PieceCounts/Sqlite/Update-32       2403696100     136937262     -94.30%
BenchmarkDB_PieceCounts/Sqlite/All-32          16454915       17416226      +5.84%
BenchmarkDB_PieceCounts/Postgres/Update-32     562646350      149697062     -73.39%
BenchmarkDB_PieceCounts/Postgres/All-32        13657257       14630751      +7.13%
```

PS: Each commit is separately reviewable.

Please describe the tests:
 - Test 1:
 - Test 2:
 
Please describe the performance impact:

## Code Review Checklist (to be filled out by reviewer)
 - [x] Does the PR describe what changes are being made?
 - [x] Does the PR describe why the changes are being made?
 - [x] Does the code follow [our style guide](https://github.com/storj/docs/blob/master/code/Style.md)?
 - [x] Does the code follow [our testing guide](https://github.com/storj/docs/blob/master/code/Testing.md)?
 - [x] Is the PR appropriately sized? (If it could be broken into smaller PRs it should be)
 - [x] Does the new code have enough tests? (*every* PR should have tests or justification otherwise. Bug-fix PRs especially)
 - [x] Does the new code have enough documentation that answers "how do I use it?" and "what does it do?"? (both source documentation and [higher level](https://github.com/storj/docs), diagrams?)
 - [x] Does any documentation need updating?
 - [ ] Do the database access patterns make sense?