@egonelbre 2019-09-11
@stefanbenten this is not for metrics, this number is used for calculating garbage collection bloom filter sizes.